Use bpool guid to identify correct bpool
Since bpool is a standard name, the system may be used with other drives
that contain another pool named bpool.  Thus the
zfs-import-bpool.service would fail since it couldn't choose between
multiple pools with the same name.

When we create the bpool we grab the guid from it.  This is used later
in setup.sh in the chroot when we create the zfs-import-bpool.service
files.  Kinda ugly in that it's hard-coded in, but it ensure that the
service will only try to import the correct bpool