Add zfs encrypted dataset support for root and /home
The main root dataset is encrypted via passphrase that is prompted at
boot.  The /home dataset uses a keyfile stored in /root/pool.key for
encryption.  This is safe because the pool.key file is not accessible
until AFTER the root dataset is unlocked via passphrase.

A zfs-multi-mount.sh script is used for unlocking datasets, triggered by
a systemd unit zfs-load-key.service.  It only unlocks datasets whose key
status is "unavailable".