Updated HereDoc file creation to use tabs for indentation
Using the format cat > filename <<- EOF  (note <<-) the heredoc will
strip any leading TAB chars from the text - not spaces. So the first
char should be a TAB (not 4 spaces) which will be removed, and any
following leading spaces will REMAIN, so the text will be properly
indented.

This allows for HereDoc to be indented in a if [ ] block which makes
reading the code much easier.