Fix HIBERNATE for LUKS encryption
Need to encrypt the swap partition FIRST when using Hibernate with LUKS
encryption.  We create a derived key from that to be able to open up all
the other disks/partitions.

Reason swap has to be first, is that it has to be unlocked before a
resume can happen to come out of hibernation ...