Incorporate proper syslinux.cfg generation for zfsbootmenu
New facility in zfsbootmenu to correctly generate a syslinux.cfg file
and to include snippets for other things.

This creates a snippet for memtest86+, so we move the install of
memtest86 further up to keep it all together with zfsbootmenu.  Meant
that unzip needed to be installed sooner as well.