Fix timeouts when booting with no network
With dracut rd.neednet=1 it will wait forever for a network to appear.
With laptops in a new location, that might be only the wlan which can't
connect to a local wifi (yet) or a docking station ethernet that isn't
connected.

Network is only really needed for Dropbear/ssh unlocking at zbm time, so
disable needing network unless Dropbear is enabled.  Mostly useful for
servers etc. that have unchanging hardware and are on network.