Fixed refind on raid boot disks, updated startup.nsh
Looks like refind-install can NOT handle a raid setup for the EFI
partitions.  It installs correctly to /boot/efi but cannot use
efibootmgr to set the EFI partitions in the UEFI nvram as boot options.

So for a raid setup, we manually use efibootmgr to set those partitions
up.

The startup.nsh script was short, dumb and blind.  The updated one
actually scans all drives for a valid refind_x64.efi and runs it when
found.