Complete transition to function modularity and add-new-dataset
The whole script has been refactored into functions for all the
working bits.  The mainline then calls the functions as required.

Aside from more neatness, this was driven by the need to add the
ability to build a new ROOT dataset for a different or parallel
Ubuntu distro.

If the current system is noble/24.04 under <pool>/ROOT/noble
this will allow building a plucky/25.04 under <pool>/ROOT/plucky
The new plucky root dataset is selectable under zfsbootmenu for
booting, and cleanly mounts /boot/efi, /root and the home user dir